requirements.txt
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
groq>=0.12.0
tavily-python>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.32.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
